Split out CampaignsTable into a standalone component.
